requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
reportlab==4.1.0
resend==2.23.0
pdfplumber==0.11.4
anthropic==0.49.0
python-dateutil==2.9.0
